+++ CHANGELOG.markdown
@@ -12,6 +12,7 @@ Version 1.2.2
 * Fixed bug in vertical `spacing` calculation on iOS (GH-188, GH-189).
 * New alternative Javascript-less initialization with HTML data attributes (GH-193).
 * Revised examples with syntax-highlighted code.
+* New AMD & CommonJS compatibility.
 
 
 Version 1.2.1